refactor: Update roomStatus object to include timestamp
This commit updates the `roomStatus` object in the code to include a `timestamp` property. The `timestamp` will be set to the current date and time whenever the room status changes. This change provides more information about when the room was locked or unlocked, improving the accuracy of the status information.

Note: The commit message has been generated based on the provided code changes and recent repository commits.