refactor: Add insertUser function to user.js
This commit adds the insertUser function to the user.js file. The function is responsible for inserting a new user into the database. It takes a student ID and password as parameters, hashes the password using bcrypt, and inserts the student ID and hashed password into the 'students' table. This change enhances the functionality of the application by allowing new users to be added to the database.