refactor: Update upload_port in platformio.ini
This commit updates the upload_port configuration in the platformio.ini file. The previous value was set to '/dev/ttyUSB0', and it has been changed to '/dev/ttyUSB*' to allow for dynamic selection of the upload port. This change improves the flexibility of the application by automatically detecting the correct upload port for the Arduino board.

Note: The commit message has been generated based on the provided code changes and recent repository commits.