refactor: Add useRoomContext hook to SignIn component
This commit adds the useRoomContext hook to the SignIn component in SignIn.jsx. The useRoomContext hook is used to access the setStudentId function from the RoomContext, allowing the student ID to be set after a successful login. This change enhances the functionality of the application by ensuring that the student ID is properly set and used throughout the application.