refactor: Update insertUser function to include user's name in database insertion
This commit modifies the insertUser function in user.js to include the user's name in the database insertion. Previously, only the student ID and password hash were inserted. This change enhances the functionality of the user registration feature by storing the user's name along with other details in the database.